# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "description": "Error is a short, human-readable description of what went\nwrong. The server never returns raw stack traces or DB\nerror strings here; the goal is to be informative without\nleaking internals.",
                        "example": "Invalid request payload",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "model.Notification": {
                "properties": {
                    "content": {
                        "description": "Notification content",
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "is_read": {
                        "description": "Whether it has been read",
                        "type": "boolean"
                    },
                    "related_id": {
                        "description": "Related resource ID",
                        "type": "string"
                    },
                    "related_post_id": {
                        "description": "Owning post ID for reply/comment notifications",
                        "type": "integer"
                    },
                    "related_type": {
                        "$ref": "#/components/schemas/model.NotificationRelatedType"
                    },
                    "title": {
                        "description": "Notification title",
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/model.NotificationType"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "user_id": {
                        "description": "Receiving user ID",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "model.NotificationRelatedType": {
                "description": "Related resource type",
                "enum": [
                    "post",
                    "comment",
                    "creator_application"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "NotificationRelatedTypePost",
                    "NotificationRelatedTypeComment",
                    "NotificationRelatedTypeCreatorApplication"
                ]
            },
            "model.NotificationType": {
                "description": "Notification type: comment, like, reply, review, role",
                "enum": [
                    "role",
                    "comment",
                    "like",
                    "reply",
                    "review"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "NotificationTypeRole",
                    "NotificationTypeComment",
                    "NotificationTypeLike",
                    "NotificationTypeReply",
                    "NotificationTypeReview"
                ]
            },
...
            "notification.MessageResponse": {
                "properties": {
                    "message": {
                        "example": "Notification marked as read",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "notification.NotificationListResponse": {
                "properties": {
                    "notifications": {
                        "items": {
                            "$ref": "#/components/schemas/model.Notification"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/notification.Pagination"
                    }
                },
                "type": "object"
            },
            "notification.Pagination": {
                "properties": {
                    "limit": {
                        "example": 10,
                        "type": "integer"
                    },
                    "page": {
                        "example": 1,
                        "type": "integer"
                    },
                    "total": {
                        "example": 42,
                        "type": "integer"
                    },
                    "totalPages": {
                        "example": 5,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "notification.UnreadCountResponse": {
                "properties": {
                    "unreadCount": {
                        "example": 3,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "JWT bearer token issued by /api/auth/login",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "contact": {
            "name": "GitHub Issues",
            "url": "https://github.com/vexgo-org/vexgo/issues"
        },
        "description": "Self-hosted blog CMS HTTP API.",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "termsOfService": "https://github.com/vexgo-org/vexgo",
        "title": "VexGo API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/notifications": {
            "get": {
                "description": "Returns the authenticated user's notifications, paginated\nand optionally filtered by type (comment, like, reply,\nreview, role) or read state.",
                "parameters": [
                    {
                        "description": "page number (1-based)",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 1,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "page size",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 10,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "notification type filter",
                        "in": "query",
                        "name": "type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "read state filter (true/false)",
                        "in": "query",
                        "name": "is_read",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/notification.NotificationListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List current user's notifications",
                "tags": [
                    "notifications"
                ]
            }
        },
        "/notifications/read-all": {
            "post": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/notification.MessageResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Mark all notifications as read",
                "tags": [
                    "notifications"
                ]
            }
        },
        "/notifications/unread-count": {
            "get": {
                "description": "Returns the number of unread notifications for the\nauthenticated user. Used by the frontend header to\nrender the badge.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/notification.UnreadCountResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Unread notification count",
                "tags": [
                    "notifications"
                ]
            }
        },
        "/notifications/{id}": {
            "delete": {
                "parameters": [
                    {
                        "description": "notification id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/notification.MessageResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "invalid id"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "notification not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete a notification",
                "tags": [
                    "notifications"
                ]
            },
            "put": {
                "parameters": [
                    {
                        "description": "notification id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/notification.MessageResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "invalid id"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "notification not found or not updated"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Mark a single notification as read",
                "tags": [
                    "notifications"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
...
}